+++ apps/balancer-tools/src/app/apr/(components)/KPI.tsx
