+++ apps/balancer-tools/src/app/apr/(components)/OverviewCards.tsx
@@ -1,4 +1,4 @@
-import { KPI } from "#/components/KPI";
+import { KPI } from "./KPI";
 
 interface CardDetail {
   title: string;
